% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function1.R
\name{CCT}
\alias{CCT}
\title{The aggregated Cauchy association test (ACAT)}
\usage{
CCT(pvals, weights = NULL)
}
\arguments{
\item{pvals}{a numeric vector of p-values, where each of the element is between 0 to 1, to be combined.}

\item{weights}{weights a numeric vector of non-negative weights. If \code{NULL}, the equal weights are assumed.}
}
\value{
return the aggregated p-value combining p-values from the vector \code{pvals}
}
\description{
The aggregated Cauchy association test (ACAT)
}
