% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function2.R
\name{Multi_Analysis}
\alias{Multi_Analysis}
\title{Calculate the P-values of the Generalized Berk-Jones (GBJ) test, the generalized higher criticism (GHC) test, the the MinimumP second-level (MinP) test and the omnibus (OMNI) test.}
\usage{
Multi_Analysis(Z_score, Sigma)
}
\arguments{
\item{Z_score}{Vector of test statistics for each factor in the set (i.e. marginal test statistic for each trait).}

\item{Sigma}{d*d matrix of the correlations between all the test statistics in the set, where d is the total number of test statistics in the set. You only need to specify EITHER cor_mat OR pairwise_cors.}
}
\value{
The p-values of the GBJ, GHC, MinP and OMNI tests, given the size of the set
}
\description{
Calculate the P-values of the Generalized Berk-Jones (GBJ) test, the generalized higher criticism (GHC) test, the the MinimumP second-level (MinP) test and the omnibus (OMNI) test.
}
